backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==5.0.5
    # via kombu
apipkg==1.5
    # via execnet
appdirs==1.4.4
    # via black
asgiref==3.3.1
    # via django
attrs==20.3.0
    # via pytest
billiard==3.6.3.0
    # via celery
black==20.8b1
    # via -r requirements.in
block-disposable-email==2.0.0
    # via django-disposable-email-checker
boto3==1.17.0
    # via -r requirements.in
boto==2.49.0
    # via -r requirements.in
botocore==1.20.0
    # via
    #   boto3
    #   s3transfer
celery==5.0.5
    # via -r requirements.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
coverage==5.4
    # via pytest-cov
django-cors-headers==3.7.0
    # via -r requirements.in
django-disposable-email-checker==2.0.2
    # via -r requirements.in
django-extensions==3.1.0
    # via -r requirements.in
django-health-check==3.16.2
    # via -r requirements.in
django-redis==4.12.1
    # via -r requirements.in
django-storages==1.11.1
    # via -r requirements.in
django==3.1.12
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-health-check
    #   django-redis
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework-camel-case==1.2.0
    # via -r requirements.in
djangorestframework-simplejwt==4.6.0
    # via -r requirements.in
djangorestframework==3.12.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
execnet==1.8.0
    # via pytest-xdist
flake8-polyfill==1.0.2
    # via pep8-naming
flake8==3.8.4
    # via
    #   -r requirements.in
    #   flake8-polyfill
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.7.0
    # via -r requirements.in
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
kombu==5.0.2
    # via celery
mccabe==0.6.1
    # via
    #   -r requirements.in
    #   flake8
mypy-extensions==0.4.3
    # via black
packaging==20.9
    # via
    #   pytest
    #   pytest-sugar
pathspec==0.8.1
    # via black
pep8-naming==0.11.1
    # via -r requirements.in
pip-tools==5.5.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.17
    # via click-repl
psycopg2-binary==2.8.6
    # via -r requirements.in
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pycodestyle==2.6.0
    # via flake8
pydocstyle==5.1.1
    # via -r requirements.in
pyflakes==2.2.0
    # via flake8
pyjwt==2.0.1
    # via djangorestframework-simplejwt
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements.in
pytest-django==4.1.0
    # via -r requirements.in
pytest-env==0.6.2
    # via -r requirements.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-pythonpath==0.7.3
    # via -r requirements.in
pytest-sugar==0.9.4
    # via -r requirements.in
pytest-xdist==2.2.0
    # via -r requirements.in
pytest==6.2.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-env
    #   pytest-forked
    #   pytest-pythonpath
    #   pytest-sugar
    #   pytest-xdist
python-dateutil==2.8.1
    # via
    #   -r requirements.in
    #   botocore
pytz==2021.1
    # via
    #   celery
    #   django
raven==6.10.0
    # via -r requirements.in
redis==3.5.3
    # via django-redis
regex==2020.11.13
    # via black
requests==2.25.1
    # via responses
responses==0.12.1
    # via -r requirements.in
s3transfer==0.3.4
    # via boto3
six==1.15.0
    # via
    #   click-repl
    #   python-dateutil
    #   responses
snowballstemmer==2.1.0
    # via pydocstyle
sqlparse==0.4.1
    # via django
termcolor==1.1.0
    # via pytest-sugar
toml==0.10.2
    # via
    #   black
    #   pytest
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via black
urllib3==1.26.5
    # via
    #   botocore
    #   requests
    #   responses
uwsgi==2.0.19.1
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==5.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
